Refactor UpdateSqlGenerator to separate select logic
Closes #10431
